Add GraphQL @defer directive support to storefront client (#2993)
* Add defer support

Add defer support by just checking if the query includes the `@defer` directive.
If it does, use the shopify graphql client. Using this client is awkward though,
because it doesn't expose the raw Resposne object. Perhaps we could refactor
everything to use it?

Also change the default `updateDiscountCodes` fragment to use the `@defer` fragment

* Delete defer test route

* Add changeset

---------

Co-authored-by: Juan P. Prieto <jp@calltheguys.co>